Move client related functions to own file
