+++ internal/utils/miscellaneous.go
